+++ src/settle.d.ts
@@ -1,0 +1,4 @@
+declare module 'axios/lib/core/settle' {
+  const settle: (resolve: (res: any) => void, reject: (reason: any) => void, res: any) => void
+  export = settle
+}